requirements.txt
numpy==1.23.2
python-dotenv==0.21.1
Flask==2.2.2
hyperdb-python==0.1.2
pandas==2.0.0
openai==0.27.4